fix: make web build work by removing dart:io from web code paths
The web build crashed at startup because getApplicationCacheDirectory()
(path_provider) has no web implementation and Cache used dart:io
File/Directory which aren't available on web.

- Make Cache abstract and platform-agnostic (no dart:io)
- Add FileCache (native, file-system-based) and WebCache (HTTP-only)
- Use conditional imports for cache creation and image display
- On web, images use Image.network() with CDN URLs directly
