+++ logic/lib/file_cache.dart
@@ -1,0 +1,6 @@
+/// Native file-system-based cache implementation.
+///
+/// Only available on platforms that support dart:io.
+library;
+
+export 'src/data/file_cache.dart';
