+++ logic/lib/web_cache.dart
@@ -1,0 +1,4 @@
+/// Web cache implementation that fetches from CDN without file-system caching.
+library;
+
+export 'src/data/web_cache.dart';
